cleaned up Custom Path Manager load/unload
