+++ TLM/TLM/LoadingExtension.cs
@@ -25,13 +25,6 @@ namespace TrafficManager {
             RegisterCustomManagers();
         }
 
-        static FastList<ISimulationManager> GetSimManagers() =>
-            typeof(SimulationManager)
-            .GetField("m_managers", BindingFlags.Static | BindingFlags.NonPublic)
-            ?.GetValue(null)
-            as FastList<ISimulationManager>
-            ?? throw new Exception("could not get SimulationManager.m_managers");
-
         internal static AppMode? AppMode => SimulationManager.instance.m_ManagersWrapper.loading.currentMode;
 
         public static SimulationManager.UpdateMode UpdateMode => SimulationManager.instance.m_metaData.m_updateMode;
@@ -43,8 +36,6 @@ namespace TrafficManager {
         /// </summary>
         internal static bool PlayMode => AppMode != null && AppMode == ICities.AppMode.Game;
 
-        public static CustomPathManager CustomPathManager { get; set; }
-
         /// <summary>
         /// Contains loaded languages and lookup functions for text translations
         /// </summary>
@@ -53,8 +44,6 @@ namespace TrafficManager {
         public static List<ICustomManager> RegisteredManagers { get; private set; }
 
         public static bool IsGameLoaded { get; private set; }
-
-        public static bool IsPathManagerReplaced { get; private set; }
 
         private static void RegisterCustomManagers() {
             RegisteredManagers = new List<ICustomManager>();
@@ -96,9 +85,7 @@ namespace TrafficManager {
         public static void Unload() {
             Log.Info("LoadingExtension.Unload()");
 
-            if (IsPathManagerReplaced) {
-                CustomPathManager._instance.WaitForAllPaths();
-            }
+            CustomPathManager._instance?.OnLevelUnloading();
 
             try {
                 foreach (ICustomManager manager in RegisteredManagers.AsEnumerable().Reverse()) {
@@ -214,72 +201,7 @@ namespace TrafficManager {
 
             IsGameLoaded = true;
 
-            //it will replace stock PathManager or already Replaced before HotReload
-            if (!IsPathManagerReplaced || TrafficManagerMod.InGameHotReload) {
-                try {
-                    Log.Info("Pathfinder Compatible. Setting up CustomPathManager and SimManager.");
-                    FieldInfo pathManagerInstance = typeof(Singleton<PathManager>).GetField(
-                        "sInstance",
-                        BindingFlags.Static | BindingFlags.NonPublic);
-                    if (pathManagerInstance == null) {
-                        throw new Exception("pathManagerInstance is null");
-                    }
-
-                    PathManager stockPathManager = PathManager.instance;
-                    if (stockPathManager == null) {
-                        throw new Exception("stockPathManager is null");
-                    }
-
-                    Log._Debug($"Got stock PathManager instance {stockPathManager?.GetName()}");
-
-                    CustomPathManager =
-                        stockPathManager.gameObject.AddComponent<CustomPathManager>();
-                    Log._Debug("Added CustomPathManager to gameObject List");
-
-                    if (CustomPathManager == null) {
-                        Log.Error("CustomPathManager null. Error creating it.");
-                        return;
-                    }
-
-                    CustomPathManager.UpdateWithPathManagerValues(stockPathManager);
-                    Log._Debug("UpdateWithPathManagerValues success");
-
-                    pathManagerInstance.SetValue(null, CustomPathManager);
-
-                    Log._Debug("Getting Current SimulationManagers");
-                    var simManagers = GetSimManagers();
-
-                    Log._Debug("Removing Stock PathManager");
-                    simManagers.Remove(stockPathManager);
-
-                    Log._Debug("Adding Custom PathManager");
-                    simManagers.Add(CustomPathManager);
-
-                    Object.Destroy(stockPathManager, 10f);
-
-                    Log._Debug("Should be custom: " + Singleton<PathManager>.instance.GetType());
-
-                    IsPathManagerReplaced = true;
-                }
-                catch (Exception ex) {
-                    string error =
-                        "Traffic Manager: President Edition failed to load. You can continue " +
-                        "playing but it's NOT recommended. Traffic Manager will not work as expected.";
-                    Log.Error(error);
-                    Log.Error($"Path manager replacement error: {ex}");
-
-                    Singleton<SimulationManager>.instance.m_ThreadingWrapper.QueueMainThread(
-                        () => {
-                            UIView.library
-                                  .ShowModal<ExceptionPanel>(
-                                      "ExceptionPanel")
-                                  .SetMessage(
-                                      "TM:PE failed to load",
-                                      error,
-                                      true);
-                        });
-                }
-            }
+            CustomPathManager.OnLevelLoaded();
 
             ModUI.OnLevelLoaded();
             if (PlayMode) {
